roles added to the logic for in the routes and the rest also made

+++ routes/api.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\CategoryController;
 use App\Http\Controllers\ImageController;
 use App\Http\Controllers\JwtAuthController;
+use App\Http\Controllers\RoleController;
 use App\Http\Controllers\StickerController;
 use Illuminate\Support\Facades\Route;
 
@@ -44,20 +45,20 @@ Route::prefix('stickers')->group(
         // Route to get a specific sticker by ID
         Route::get('/{id}', [StickerController::class, 'show']);
         // Route to create a new sticker
-        Route::post('/', [StickerController::class, 'store'])->middleware('jwt_full');
+        Route::post('/', [StickerController::class, 'store'])->middleware(['jwt_full', 'admin']);
         // Route to update an existing sticker by ID
-        Route::put('/{id}', [StickerController::class, 'update'])->middleware('jwt_full');
+        Route::put('/{id}', [StickerController::class, 'update'])->middleware(['jwt_full', 'admin']);
         // Route to delete an existing sticker by ID
-        Route::delete('/{id}', [StickerController::class, 'destroy'])->middleware('jwt_full');
+        Route::delete('/{id}', [StickerController::class, 'destroy'])->middleware(['jwt_full', 'admin']);
 
         // Route to get a list of images associated with a specific sticker
         Route::get('/{stickerId}/images', [ImageController::class, 'index']);
         // Route to create a new image for a specific sticker
-        Route::post('/{stickerId}/images', [ImageController::class, 'store'])->middleware('jwt_full');
+        Route::post('/{stickerId}/images', [ImageController::class, 'store'])->middleware(['jwt_full', 'admin']);
         // Route to update an existing image for a specific sticker by ID
-        Route::put('/{stickerId}/images/{imageId}', [ImageController::class, 'update'])->middleware('jwt_full');
+        Route::put('/{stickerId}/images/{imageId}', [ImageController::class, 'update'])->middleware(['jwt_full', 'admin']);
         // Route to delete an existing image for a specific sticker by ID
-        Route::delete('/{stickerId}/images/{imageId}', [ImageController::class, 'destroy'])->middleware('jwt_full');
+        Route::delete('/{stickerId}/images/{imageId}', [ImageController::class, 'destroy'])->middleware(['jwt_full', 'admin']);
     }
 );
 
@@ -68,10 +69,26 @@ Route::prefix('categories')->group(
         // Route to get a specific category by ID
         Route::get('/{id}', [CategoryController::class, 'show']);
         // Route to create a new category
-        Route::post('/', [CategoryController::class, 'store'])->middleware('jwt_full');
+        Route::post('/', [CategoryController::class, 'store'])->middleware(['jwt_full', 'admin']);
         // Route to update an existing category by ID
-        Route::put('/{id}', [CategoryController::class, 'update'])->middleware('jwt_full');
+        Route::put('/{id}', [CategoryController::class, 'update'])->middleware(['jwt_full', 'admin']);
         // Route to delete an existing category by ID
-        Route::delete('/{id}', [CategoryController::class, 'destroy'])->middleware('jwt_full');
+        Route::delete('/{id}', [CategoryController::class, 'destroy'])->middleware(['jwt_full', 'admin']);
+    }
+);
+
+
+Route::prefix('roles')->group(
+    function ($router) {
+        // Route to get a list of all categories
+        Route::get('/', [RoleController::class, 'index'])->middleware(['jwt_full', 'admin']);
+        // Route to get a specific category by ID
+        Route::get('/{id}', [RoleController::class, 'show'])->middleware(['jwt_full', 'admin']);
+        // Route to create a new category
+        Route::post('/', [RoleController::class, 'store'])->middleware(['jwt_full', 'admin']);
+        // Route to update an existing category by ID
+        Route::put('/{id}', [RoleController::class, 'update'])->middleware(['jwt_full', 'admin']);
+        // Route to delete an existing category by ID
+        Route::delete('/{id}', [RoleController::class, 'destroy'])->middleware(['jwt_full', 'admin']);
     }
 );
